+++ training/tests/integration/aicon/test_cicd_aicon_04_icon-dream_medium.py
@@ -15,12 +15,11 @@ import os
 import pathlib
 from functools import reduce
 from operator import getitem
-from pathlib import Path
 
 import matplotlib as mpl
 import pytest
 import torch
-from conftest import GetTmpPaths
+from conftest import GetTmpPath
 from hydra import compose
 from hydra import initialize
 from omegaconf import DictConfig
@@ -39,19 +38,18 @@ mpl.use("agg")
 
 @pytest.fixture
 @typechecked
-def aicon_config_with_tmp_dir(get_tmp_paths: GetTmpPaths, get_test_archive: GetTestArchive) -> DictConfig:
+def aicon_config_with_tmp_dir(get_tmp_path: GetTmpPath, get_test_archive: GetTestArchive) -> DictConfig:
     """Get AICON config with temporary output paths."""
     with initialize(version_base=None, config_path="./"):
         config = compose(config_name="test_cicd_aicon_04_icon-dream_medium")
 
-    tmp_dir, rel_paths, dataset_urls = get_tmp_paths(config, ["dataset", "forcing_dataset"])
-    config.system.output.root = tmp_dir
-    dataset, forcing_dataset = rel_paths
-    config.system.input.dataset = str(Path(tmp_dir, dataset))
-    config.system.input.forcing_dataset = str(Path(tmp_dir, forcing_dataset))
+    tmp_dir_dataset, url_dataset = get_tmp_path(config.system.input.dataset)
+    tmp_dir_forcing_dataset, url_forcing_dataset = get_tmp_path(config.system.input.forcing_dataset)
+    config.system.input.dataset = str(tmp_dir_dataset)
+    config.system.input.forcing_dataset = str(tmp_dir_forcing_dataset)
 
-    for url in dataset_urls:
-        get_test_archive(url)
+    get_test_archive(url_dataset)
+    get_test_archive(url_forcing_dataset)
 
     return config